refactor: move the components to atom structure
