+++ src/components/01-atoms/Jazzicons.tsx
@@ -8,12 +8,10 @@ type JazziconsProps = {
   size?: number;
 };
 
-const Jazzicons: FC<JazziconsProps> = ({ seed, size }) => {
+export const Jazzicons: FC<JazziconsProps> = ({ seed, size }) => {
   if (!seed) return <Skeleton height={40} />;
 
   if (size) return <Jazzicon seed={jsNumberForAddress(seed)} diameter={size} />;
 
   return <Jazzicon seed={jsNumberForAddress(seed)} />;
 };
-
-export default Jazzicons;
