+++ src/components/01-atoms/index.ts
@@ -1,0 +1,6 @@
+export { AddressInput } from "./AddressInput";
+export { DarkModeButton } from "./DarkModeButton";
+export { Footer } from "./Footer";
+export { Header } from "./Header";
+export { InfoText } from "./InfoText";
+export { SignMessage } from "./SignMessage";
