+++ src/components/02-molecules/BlockNumber.tsx
@@ -2,12 +2,10 @@ import { type FC } from "react";
 
 import { useBlockNumber } from "wagmi";
 
-import { InfoText } from "@/components";
+import { InfoText } from "@/components/01-atoms";
 
-const BlockNumber: FC = (): JSX.Element => {
+export const BlockNumber: FC = (): JSX.Element => {
   const { data } = useBlockNumber({ watch: true });
 
   return <InfoText label="Block Number" value={data?.toString()} />;
 };
-
-export default BlockNumber;
