+++ src/components/02-molecules/Chain.tsx
@@ -2,14 +2,12 @@ import { type FC } from "react";
 
 import { useAccount } from "wagmi";
 
-import { InfoText } from "@/components";
+import { InfoText } from "@/components/01-atoms";
 
-const Chain: FC = (): JSX.Element => {
+export const Chain: FC = (): JSX.Element => {
   const { chain } = useAccount();
 
   const chainInfo = `${chain?.name} (${chain?.id})`;
 
   return <InfoText label="Chain" value={chainInfo} />;
 };
-
-export default Chain;
