+++ src/components/04-templates/index.ts
@@ -1,0 +1,1 @@
+export { MainPane } from "./MainPane";
