--- src/components/AddressInput/index.ts
@@ -1,1 +1,0 @@
-export { default as AddressInput } from "./AddressInput";
