--- src/components/DarkModeButton/index.ts
@@ -1,1 +1,0 @@
-export { default as DarkModeButton } from "./DarkModeButton";
