--- src/components/InfoText/index.ts
@@ -1,1 +1,0 @@
-export { default as InfoText } from "./InfoText";
