--- src/components/MainPane/components/index.ts
@@ -1,7 +1,0 @@
-export { default as Status } from "./Status";
-export { default as Address } from "./Address";
-export { default as Chain } from "./Chain";
-export { default as Balance } from "./Balance";
-export { default as BlockNumber } from "./BlockNumber";
-export { default as SignMessage } from "./SignMessage";
-export { default as TransferNative } from "./TransferNative";
