--- src/components/MainPane/index.ts
@@ -1,1 +1,0 @@
-export { default as MainPane } from "./MainPane";
