--- src/components/index.ts
@@ -1,6 +1,0 @@
-export * from "./AddressInput";
-export * from "./DarkModeButton";
-export * from "./Footer";
-export * from "./Header";
-export * from "./MainPane";
-export * from "./InfoText";
